Update workspace settings and extend terminal functionalities in Plugin.cs
This commit includes an update to the workspace settings file to change how certain properties are stored. In addition, the functionality of the terminal in Plugin.cs has been significantly extended. It includes the addition of new commands such as toggling ship lights, adding suits, scanning the base, getting the current time, and so on. Further, detailed comments have been added for better code readability and